+++ lib/Auth/Process/DatabaseCommand.php
@@ -2,6 +2,7 @@
 
 namespace SimpleSAML\Module\proxystatistics\Auth\Process;
 
+use SimpleSAML\Error\Exception;
 use SimpleSAML\Logger;
 
 /**
@@ -18,10 +19,26 @@ class DatabaseCommand
         $statisticsTableName = $databaseConnector->getStatisticsTableName();
         $identityProvidersMapTableName = $databaseConnector->getIdentityProvidersMapTableName();
         $serviceProvidersMapTableName = $databaseConnector->getServiceProvidersMapTableName();
-        $idpEntityID = $request['saml:sp:IdP'];
-        $idpName = $request['Attributes']['sourceIdPName'][0];
-        $spEntityId = $request['Destination']['entityid'];
-        $spName = $request['Destination']['name']['en'];
+
+        if ($databaseConnector->getMode() === 'IDP') {
+            $idpName = $databaseConnector->getIdpName();
+            $idpEntityID = $databaseConnector->getIdpEntityId();
+            $spEntityId = $request['Destination']['entityid'];
+            $spName = $request['Destination']['name']['en'];
+        } elseif ($databaseConnector->getMode() === 'SP') {
+            $idpName = $request['Attributes']['sourceIdPName'][0];
+            $idpEntityID = $request['saml:sp:IdP'];
+            $spEntityId = $databaseConnector->getSpEntityId();
+            $spName = $databaseConnector->getSpName();
+        } elseif ($databaseConnector->getMode() === 'PROXY') {
+            $idpName = $request['Attributes']['sourceIdPName'][0];
+            $idpEntityID = $request['saml:sp:IdP'];
+            $spEntityId = $request['Destination']['entityid'];
+            $spName = $request['Destination']['name']['en'];
+        } else {
+            throw new Exception('Unknown mode is set. Mode has to be one of the following: PROXY, IDP, SP.');
+        }
+
         $year = $date->format('Y');
         $month = $date->format('m');
         $day = $date->format('d');